Added user actions inside qmanager and fixed missing waiting call UI
